v1.0 正式发布
- 实现了最后两个属性: 生命恢复和物理防御
- 更新了帮助列表
- 随着新属性的实现添加了两条提示信息
- 更新了版本号

+++ SoulRing/src/main/java/org/serverct/sir/soulring/runnable/RegenerationTask.java
@@ -6,9 +6,9 @@ import org.bukkit.scheduler.BukkitRunnable;
 public class RegenerationTask extends BukkitRunnable {
 
     private Player target;
-    private int heal;
+    private double heal;
 
-    public RegenerationTask(Player player, int heal) {
+    public RegenerationTask(Player player, double heal) {
         this.target = player;
         this.heal = heal;
     }
@@ -16,7 +16,7 @@ public class RegenerationTask extends BukkitRunnable {
     @Override
     public void run() {
         if(target.getHealth() < target.getMaxHealth()) {
-            target.setHealth(target.getHealth() + (double) heal);
+            target.setHealth(target.getHealth() + heal);
         }
     }
 }
